Update README.md
The momentum project is on the back burner at least for the present.  I am leaving the project up in an unfinished state and adding this as a reference to the Python,sqlite3 and statistics Project.   I have acquired about 700k lines of data.  This may be enough to verify the veracity of the gravity experiment .  (Is gravity constant?)  I am turning my attention to the Python, sqlite3 statistics project .