Update README.md
The momentum project, on a conceptual level is moving forward nicely, but the 'boots on the ground' aspect is back burner.  I am  leaving this project up sort of as a spacer, or place holder.  I have turned my attention to Python, sqlite3 and software development.
Github is cool, but this is another learning curve.  okay