Update January 6, 2022
I am remaking all three repositories with the goal of presenting this project.  Even the rotating pendulums data has interesting trends.  My thinking is thus:

there are three pendulums.  If I see weird behavior in one of the pendulums, then that is just a fluke,  But if I see this same statistically significant behavior in two pendulums then it deserves attention.