Is mass constant?  This is a boolean yes/no.
Every point is an average from one day.  A magnification of a boring stretch of value shows lots of noise and it may be correlated.  hummmm  I put data into sqlite3 with Python and then, again with Python, summarize the data into summary sqlite3 databases.  This resulted in 271 lines for this example.  I simply cut the data out of the database (I use DB Browser) and past it into Excel.  The Python platform is awkward, but powerful.  It is presently configured for one day at a time.  Each point in these graphs is an average of about 200 datum points.